+++ src/main/java/online/kingdomkeys/kingdomkeys/command/ModCommands.java
@@ -13,6 +13,7 @@ public class ModCommands {
                 .then(ChoiceCommand.register())
                 .then(DimensionCommand.register())
                 .then(DriveLevelCommand.register())
+                .then(SynthLevelCommand.register())
                 .then(DrivePointsCommand.register())
                 .then(ExpCommand.register())
                 .then(ExportWorldCommand.register())