+++ src/main/java/online/kingdomkeys/kingdomkeys/network/cts/CSChangeStyle.java
@@ -10,7 +10,7 @@ import online.kingdomkeys.kingdomkeys.KingdomKeys;
 import online.kingdomkeys.kingdomkeys.data.PlayerData;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.enums.DualChoices;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.enums.HandStyle;
-import online.kingdomkeys.kingdomkeys.integration.epicfight.EpicFightUtils;
+import online.kingdomkeys.kingdomkeys.integration.epicfight.EpicFightEvents;
 import online.kingdomkeys.kingdomkeys.integration.epicfight.enums.SingleChoices;
 import online.kingdomkeys.kingdomkeys.network.Packet;
 import online.kingdomkeys.kingdomkeys.network.PacketHandler;
@@ -38,7 +38,7 @@ public record CSChangeStyle(String style, String handStyle) implements Packet {
         PacketHandler.syncToAllAround(player, playerData);
 
         if (KingdomKeys.efmLoaded) {
-            EpicFightUtils.refreshLivingMotions(player);
+            EpicFightEvents.refreshLivingMotions(player);
         }
     }