+++ src/main/java/online/kingdomkeys/kingdomkeys/network/cts/CSUpgradeBagPacket.java
@@ -37,6 +37,9 @@ public record CSUpgradeBagPacket() implements Packet {
 		if(stack == null || stack.isEmpty()){
 			stack = Utils.getItemInAnyHand(player, ModItems.shotlocksBag.get());
 		}
+		if(stack == null || stack.isEmpty()){
+			stack = Utils.getItemInAnyHand(player, ModItems.keychainsBag.get());
+		}
 
 		if(stack != null) {
 			int bagLevel = stack.get(ModComponents.BAG_LEVEL);